pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proteinqc"
version = "0.2.0"
description = "RL/MCTS-based mRNA ORF exploration and protein quality prediction"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Gong Chen", email = "gongchen@example.com"}
]
keywords = ["RNA", "protein", "CaLM", "classification", "deep-learning"]

dependencies = [
    "torch>=2.0.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "numpy>=1.24.0",
    "biopython>=1.83",
    "python-codon-tables>=0.1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
verify = [
    "multimolecule>=0.1.0",
]
train = [
    "datasets>=2.0.0",
    "accelerate>=0.20.0",
]
deploy = [
    "coremltools>=7.0.0",
    "ane-transformers>=0.1.0",
]
agent = [
    "mlx>=0.22; sys_platform == 'darwin'",
    "mlx-lm>=0.25; sys_platform == 'darwin'",
    "psutil>=5.9.0",
]
smolagents = [
    "smolagents>=1.0.0",
    "mlx-lm>=0.25; sys_platform == 'darwin'",
]

[project.scripts]
benchmark = "proteinqc.cli.benchmark:main"
benchmark-multispecies = "proteinqc.cli.benchmark_multispecies:main"
rna-challenge-retest = "proteinqc.cli.rna_challenge_retest:main"
save-mlp-head = "proteinqc.cli.save_mlp_head:main"
bake-evidence = "proteinqc.cli.bake_evidence:main"
train-agent = "proteinqc.cli.train_agent:main"
run-agent = "proteinqc.cli.run_agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["proteinqc*"]
namespaces = false

[tool.setuptools.package-data]
"proteinqc.tools" = ["data/*.npz"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
